Create `send_transaction_test` module
Move the send transaction using lightwalletd test and its helper
functions into a new module.
